Refined QGraphicsScene implementation of GLWidget.
We still use it just as a test as it still display border artifacts.
On our side of implementation we still need to fix GLOverlay items that
act weirdly as stop GL scene stops redrawing. See beginNativePainting
and gl update(), check if the problem comes from passing the
QGraphicsScene painter to GLOverlay.

Default mode is still based on QLWidget.