QGraphicsView mode -> almost fixed margin problem.
There's a bug in Qt where a QGraphicsView as a fixed margin of 2 after
some internal method has been called (like fitInView()). We put a
negative margin for left right top, problem is that we can only put a
bottom margin of -1 or everything will start to screw up, we need -3. We
may set the back ground to the same apz color when gl scene is
maximized.

Default mode is still QLWidget.